+++ draft-packages/hierarchical-menu/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@1.2.0...@kaizen/draft-hierarchical-menu@1.2.1) (2020-08-24)
+
+**Note:** Version bump only for package @kaizen/draft-hierarchical-menu
+
+
+
+
+
+# [1.2.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@1.1.1...@kaizen/draft-hierarchical-menu@1.2.0) (2020-08-19)
+
+
+### Features
+
+* Hierarchical menu and hierarchical select improvements [PHRI-32] ([#718](https://github.com/cultureamp/kaizen-design-system/issues/718)) ([e647ffa](https://github.com/cultureamp/kaizen-design-system/commit/e647ffa4c749f9ed0996934493b640b5843ca3a4))
+
+
+
+
+
 ## [1.1.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hierarchical-menu@1.1.0...@kaizen/draft-hierarchical-menu@1.1.1) (2020-08-19)
 
 **Note:** Version bump only for package @kaizen/draft-hierarchical-menu